requirements.txt
pandas>=1.5.0
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.66.0
# Optional pretty output in reports; app falls back gracefully if missing
prettytable>=3.9.0

# Web UI stack (FastAPI + Jinja + HTMX served via CDN)
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
Jinja2>=3.1.0
python-multipart>=0.0.9

# Config/schema validation
pydantic>=2.5.0

# YAML parsing for theme whitelist governance
PyYAML>=6.0

# Development dependencies are in requirements-dev.txt